backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-dotenv==1.0.1
chromadb==0.5.3
tavily-python>=0.5.0
alembic==1.13.1
pydantic-settings==2.1.0
httpx==0.27.0
supabase==2.9.1
pyjwt==2.8.0
numpy==1.26.4

beautifulsoup4==4.12.3

tiktoken>=0.7.0
